3/requirements.txt
pytest
pytest-logger
pytest-html

pytest-selenium

pyhamcrest
jsonschema

requests
